move scrypt to own file
