+++ packages/server/src/db/util.ts
@@ -1,4 +1,3 @@
-import crypto from 'crypto'
 import { sql } from 'kysely'
 import { EntityTarget, SelectQueryBuilder } from 'typeorm'
 
@@ -58,27 +57,4 @@ export const existsByCreatorSubquery = (
       .from(table, 'table')
       .where('table.creator = :creator', { creator })
   }
-}
-
-export const scryptHash = (password: string): Promise<string> => {
-  return new Promise((resolve, reject) => {
-    const salt = crypto.randomBytes(16).toString('hex')
-    crypto.scrypt(password, salt, 64, (err, hash) => {
-      if (err) reject(err)
-      resolve(salt + ':' + hash.toString('hex'))
-    })
-  })
-}
-
-export const scryptVerify = (
-  password: string,
-  storedHash: string,
-): Promise<boolean> => {
-  return new Promise((resolve, reject) => {
-    const [salt, hash] = storedHash.split(':')
-    crypto.scrypt(password, salt, 64, (err, derivedHash) => {
-      if (err) reject(err)
-      resolve(hash === derivedHash.toString('hex'))
-    })
-  })
 }
